requirements.txt
python-chess
pandas
dash>2